Move RelayChainCli related logic under cli.rs
